Ctrl crash, compiler warning and excessive log related changes
This commit contains changes related to
a) Removing of assert if dm is not found to ignore and proceed without
crashing so as not to impact normal operation.
b) Addressing few compiler warnings in controller.
c) Remove excessive logs of unhandled message in find_em_for_msg_type.

+++ src/ctrl/em_cmd_ctrl.cpp
@@ -136,7 +136,7 @@ int em_cmd_ctrl_t::send_result(em_cmd_out_status_t status)
 
     tmp = reinterpret_cast<unsigned char *> (m_cmd.status_to_string(status, str));
 
-    if ((ret = SSL_write(m_ssl, tmp, strlen(str) + 1)) <= 0) {
+    if ((ret = SSL_write(m_ssl, tmp, static_cast<int> (strlen(str) + 1))) <= 0) {
         printf("%s:%d: write error on socket, err:%d\n", __func__, __LINE__, errno);
     }
